% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileB2D.R
\name{compileB2D}
\alias{compileB2D}
\title{Compile bookdown formatted markdown to a Dodona-ready format}
\usage{
compileB2D(
  source_dir = ".",
  output_dir = "./dodona_md",
  book_dir = "book_md",
  language = "nl",
  split_level = 2,
  continue_str = "Vervolg",
  config_file = "_bookdown.yml"
)
}
\arguments{
\item{source_dir}{(optional) Path to the directory containing the bookdown source files. Default is current working directory.}

\item{output_dir}{(optional) Path to the directory where the Dodona file structure will be saved. Default is './dodona_md'}

\item{book_dir}{(optional) Name of the directory where the complete book will be saved as an .md file. This directory will be placed in the source directory. Default is 'book_md'}

\item{language}{(optional) Language identifier string that is used to tell Dodona which language the descriptions are in. Default is 'nl'.}

\item{split_level}{(optional) Level of markdown headers at which chapters will be split in Dodona reading assignments. Default is 2, corresponding to '##' headers.}

\item{continue_str}{(optional) String that is used to indicate the continuation of a chapter when section breaks (<!--break-->) are used. Default is 'Vervolg'.}

\item{config_file}{(optional) Name of the configuration file used by bookdown. Default is '_bookdown.yml'}
}
\description{
The compileB2D function can be used to compile a set of Rmarkdown files
that follow the bookdown standard to a folder structure of description files that can be directly used with Dodona.
}
\examples{
#compileB2D()
#compileB2D('./myBookdownProject', output_dir='myDodonaBook', book_dir='myFullBook')

}
